chore: rename headshots
