+++ src/images/headshot/index.ts
@@ -1,11 +1,12 @@
 export { default as Amongus } from './amongus.webp'
-export { default as Christmas } from './cartoon-christmas.webp'
-export { default as Confused } from './cartoon-confused.webp'
-export { default as Cyberpunk } from './cartoon-cyberpunk.webp'
-export { default as Grumpy } from './cartoon-grumpy.webp'
-export { default as Halloween } from './cartoon-halloween.webp'
-export { default as Headshot } from './cartoon-headshot.webp'
-export { default as Ladybird } from './cartoon-ladybird.webp'
-export { default as Matrix } from './cartoon-matrix.webp'
-export { default as Thanksgiving } from './cartoon-thanksgiving.webp'
+export { default as Birthday } from './birthday.webp'
+export { default as Christmas } from './christmas.webp'
+export { default as Confused } from './confused.webp'
+export { default as Cyberpunk } from './cyberpunk.webp'
+export { default as Grumpy } from './grumpy.webp'
+export { default as Halloween } from './halloween.webp'
+export { default as Headshot } from './headshot.webp'
+export { default as Ladybird } from './ladybird.webp'
+export { default as Matrix } from './matrix.webp'
 export { default as Pride } from './pride.webp'
+export { default as Thanksgiving } from './thanksgiving.webp'
